Updated based on feedback.
